Fix rectangles: add dummy phase in canonicalize
